fix: Issue with behavior of clicking tray on mac
Clicking on tray icon hides window and so when clicking
the dock icon when window is hidden doesn't show the window
(window remains hidden). Added the 'activate' event to track
user dock icon click. Also, when window is opened but not
focused, clicking on tray hides the window. As user, we
expect the window to show. Upgrading the click logic.